Allow overriding of include/configs/board.h through CUSTOM_CONFIG environment variable.
To use this, specify a CUSTOM_CONFIG in your call to make xyz_config like so:

     $ make CUSTOM_CONFIG=~/my_custom_config.h ARCH=arm CROSS_COMPILE=arm-linux- board_config

The custom configuration file will be #included __after__ the board's configuration to allow
overrides of things like CONFIG_BOOTCOMMAND, CONFIG_BOOTDELAY, and CONFIG_EXTRA_ENV that may
be application-specific.

The purpose of this is to not pollute the git history of the board in order to override one
or two settings. If things get complicated, use a branch instead.